fix: azurite and refactoring
