+++ docker-compose.yml
@@ -41,5 +41,29 @@ services:
       - "8025:8025"
       - "1025:1025"
 
+  azurite:
+    image: mcr.microsoft.com/azure-storage/azurite
+    command: azurite --blobHost 0.0.0.0 --skipApiVersionCheck
+    ports:
+      - "10000:10000"
+    volumes:
+      - azurite-data:/data
+    extra_hosts:
+      - "localhost:host-gateway"
+    networks:
+      - portabase
+    #  DefaultEndpointsProtocol=http;
+    #  AccountName=devstoreaccount1;
+    #  AccountKey=Eby8vdM02xNOcqFlqUwJPLlmEtlCDXJ1OUzFT50uSRZ6IFsuFq2UVErCz4I6tq/K1SZFPTOtr/KBHBeksoGMGw==;
+    #  BlobEndpoint=http://127.0.0.1:10000/devstoreaccount1;
+    #  QueueEndpoint=http://127.0.0.1:10001/devstoreaccount1;
+    #  TableEndpoint=http://127.0.0.1:10002/devstoreaccount1;
+
 volumes:
   postgres-data:
+  azurite-data:
+
+networks:
+  portabase:
+    name: portabase_network
+    external: true
