+++ src/features/channel/storages/index.ts
@@ -6,7 +6,7 @@ import {
 
 import {uploadLocal, getLocal, deleteLocal, pingLocal, copyLocal} from './local';
 import {copyS3, deleteS3, getS3, pingS3, uploadS3} from "@/features/channel/storages/s3";
-import {copyBlob, deleteBlob, getBlob, pingBlob, uploadBlob} from "@/features/channel/storages/Blob";
+import {copyBlob, deleteBlob, getBlob, pingBlob, uploadBlob} from "@/features/channel/storages/az-blob";
 import {
     copyGoogleDrive,
     deleteGoogleDrive,
